pyproject.toml
[tool.poetry]
name = "fortune-api"
version = "0.1.0"
description = ""
authors = ["CamiaCL <camilaclary@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.1.3"
python-dotenv = "^0.20.0"
marshmallow-sqlalchemy = "^0.28.1"
connexion = {extras = ["swagger-ui"], version = "^2.14.0"}
Flask-Assets = "^2.0"
Flask-Injector = "^0.14.0"
pyScss = "^1.4.0"
SQLAlchemy = "^1.4.39"
Werkzeug = "^2.2.0"
jwt = "^1.3.1"
PyJWT = "^2.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
